[PIDM-873] fix: moving monetary serializer in common package
