+++ src/main/java/it/gov/pagopa/fdr/controller/model/flow/response/SingleFlowCreatedResponse.java
@@ -3,12 +3,12 @@ package it.gov.pagopa.fdr.controller.model.flow.response;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import it.gov.pagopa.fdr.controller.middleware.parser.ISO8601LocalDateSerializer;
+import it.gov.pagopa.fdr.controller.middleware.serialization.ISO8601LocalDateSerializer;
+import it.gov.pagopa.fdr.controller.middleware.serialization.MonetarySerializer;
 import it.gov.pagopa.fdr.controller.model.flow.Receiver;
 import it.gov.pagopa.fdr.controller.model.flow.Sender;
 import it.gov.pagopa.fdr.controller.model.flow.enums.ReportingFlowStatusEnum;
 import it.gov.pagopa.fdr.util.constant.ControllerConstants;
-import it.gov.pagopa.fdr.util.serialization.MonetarySerializer;
 import java.time.Instant;
 import java.time.LocalDate;
 import lombok.Builder;
